requirements.txt
python-dotenv==1.0.0
langchain==0.0.231
openai==0.27.8